Externalize date time format provider
